requirements.txt
flake8
pycodestyle
mypy
ruff
pylint
PyYAML~=6.0.2